fix: 인터페이스 타입 파일 제거
